requirements.txt
google-generativeai>=0.3.0
pathlib
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
python-dotenv>=1.0.0
